refactored file loading and shuffling. reduced number of model parameters by several orders of magnitude. model saves as two components, json and hdf5

+++ utils/utils.py
@@ -16,7 +16,9 @@ import nibabel as nib
 import sys
 from datetime import datetime
 from keras.utils import to_categorical
+from sklearn.utils import shuffle
 
+os.environ['FSLOUTPUTTYPE'] = 'NIFTI_GZ'
 
 def parse_args(session):
     '''
@@ -33,40 +35,50 @@ def parse_args(session):
     if session == "train":
         parser.add_argument('--datadir', required=True, action='store', dest='TRAIN_DIR',
                             help='Where the initial unprocessed data is')
-        parser.add_argument('--o', required=True, action='store', dest='OUT_DIR',
+        parser.add_argument('--weightdir', required=True, action='store', dest='OUT_DIR',
                             help='Output directory where the trained models are written')
+        parser.add_argument('--numcores', required=False, action='store', dest='numcores',
+                            default='1', type=int,
+                            help='Number of cores to preprocess in parallel with')
     elif session == "test":
         parser.add_argument('--infile', required=True, action='store', dest='INFILE',
                             help='Image to classify')
         parser.add_argument('--model', required=True, action='store', dest='model',
-                            help='Learnt model (.hdf5) file')
-        parser.add_argument('--o', required=True, action='store', dest='OUTFILE',
-                            help='Output filepath and name to where the results are written')
-        parser.add_argument('--preprocesseddir', required=True, action='store',
-                            dest='PREPROCESSED_DIR',
-                            help='Output directory where final preprocessed images are placed ')
+                            help='Model Architecture (.json) file')
+        parser.add_argument('--weights', required=True, action='store', dest='weights',
+                            help='Learnt weights (.hdf5) file')
+        parser.add_argument('--result_dst', required=True, action='store', dest='OUTFILE',
+                            help='Output filename (e.g. result.txt) to where the results are written')
+        # parser.add_argument('--preprocesseddir', required=True, action='store',
+        #                    dest='PREPROCESSED_DIR',
+        #                    help='Output directory where final preprocessed images are placed ')
     elif session == "validate":
+        parser.add_argument('--numcores', required=False, action='store', dest='numcores',
+                            default='1', type=int,
+                            help='Number of cores to preprocess in parallel with')
         parser.add_argument('--datadir', required=True, action='store', dest='VAL_DIR',
                             help='Where the initial unprocessed data is')
         parser.add_argument('--model', required=True, action='store', dest='model',
-                            help='Learnt model (.hdf5) file')
-        parser.add_argument('--o', required=True, action='store', dest='OUTFILE',
+                            help='Model Architecture (.json) file')
+        parser.add_argument('--weights', required=True, action='store', dest='weights',
+                            help='Learnt weights (.hdf5) file')
+        parser.add_argument('--result_dst', required=True, action='store', dest='OUTFILE',
                             help='Output directory where the results are written')
     else:
         print("Invalid session. Must be one of \"train\", \"validate\", or \"test\"")
         sys.exit()
 
-    parser.add_argument('--numcores', required=False, action='store', dest='numcores',
-                        default='1', type=int,
-                        help='Number of cores to preprocess in parallel with')
     parser.add_argument('--task', required=True, action='store', dest='task',
-                        help='Type of task: modality, T1-contrast, FL-contrast')
+                        help='Type of task: modality, T1-contrast, FL-contrast, '
+                             'modality indicates classification between several pre-contrast images, such as '
+                             'T1 vs T2 vs FLAIR. T1-contrast indicates classification between pre-contrast T1 '
+                             'and post-contrast T1. Similar comparison is for FL-contrast.')
     parser.add_argument('--gpuid', required=False, action='store', type=int, dest='GPUID',
                         help='For a multi-GPU system, the trainng can be run on different GPUs.'
-                        'Use a GPU id (single number), eg: 0 or 1 to run on that particular GPU.'
-                        '0 indicates first GPU.  Optional argument. Default is the last GPU.')
+                        'Use a GPU id (single number), eg: 1 or 2 to run on that particular GPU.'
+                        '0 indicates first GPU.  Optional argument. Default is the first GPU.')
     parser.add_argument('--delete_preprocessed_dir', required=False, action='store', dest='clear',
-                        default='n', help='delete tmp directory')
+                        default='n', help='delete all temporary directories. Enter either y or n. Default is n.')
 
     return parser.parse_args()
 
@@ -93,22 +105,27 @@ def preprocess(filename, outdir, tmpdir, reorient_script_path, robustfov_script_
     basename = os.path.basename(filename)
 
     # convert image to 256^3 1mm^3 coronal images with intensity range [0,255]
-    call = "mri_convert -c" + " " + filename + \
+    call = "mri_convert -odt uchar --crop 0 0 0 -c" + " " + filename + \
         " " + os.path.join(tmpdir, basename)
     if verbose == 0:
         call = call + " " + ">/dev/null"
     os.system(call)
 
     # reorient to RAI. Not necessary
-    call = reorient_script_path + " " + \
-        os.path.join(tmpdir, basename) + " " + "RAI"
+    #call = reorient_script_path + " " + \
+    #    os.path.join(tmpdir, basename) + " " + "RAI"
+    infile = os.path.join(tmpdir, basename)
+    outfile = os.path.join(tmpdir, "reorient_" + basename)
+    call = "3dresample -orient RAI -inset " + infile + " -prefix " + outfile
     if verbose == 0:
         call = call + " " + ">/dev/null"
     os.system(call)
-
+    
     # robustfov to make sure neck isn't included
-    call = robustfov_script_path + " " + os.path.join(tmpdir, basename) + " " +\
-        os.path.join(tmpdir, "robust_" + basename) + " " + "160"
+    infile = os.path.join(tmpdir, "reorient_" + basename)
+    outfile = os.path.join(tmpdir, "robust_" + basename)
+    call = robustfov_script_path + " " + infile + " " +\
+        outfile + " " + "160"
     if verbose == 0:
         call = call + " " + ">/dev/null"
     os.system(call)
@@ -116,11 +133,21 @@ def preprocess(filename, outdir, tmpdir, reorient_script_path, robustfov_script_
     # 3dWarp to make images AC-PC aligned. Not necessary.  Ideally images should be
     # rigid registered to some template for uniformity, but rigid registration is slow
     # This is a faster way.  -newgrid 2 will resample the image to 2mm^3 resolution
-    outfile = os.path.join(outdir, basename)
     infile = os.path.join(tmpdir, "robust_" + basename)
+    outfile = os.path.join(outdir, basename)
     call = "3dWarp -deoblique -NN -newgrid 2 -prefix" + " " + outfile + " " + infile
     if verbose == 0:
-        call = call + " " + ">/dev/null"
+        call = call + " " + ">/dev/null 2>&1"
+    os.system(call)
+
+    # since the intensities are already [0,255], change the file from float to uchar to save space
+    call = "fslmaths " + outfile + " " + outfile + " -odt char"
+    os.system(call)
+
+    # delete temporary files to save space, otherwise the temp directory takes more than 100GB
+    call = "rm -f " + os.path.join(tmpdir, basename)
+    os.system(call)
+    call = "rm -f " + os.path.join(tmpdir, "robust_" + basename)
     os.system(call)
 
     return os.listdir(outdir)[0]
@@ -136,9 +163,8 @@ def preprocess_dir(train_dir, preprocess_dir, reorient_script_path, robustfov_sc
         - reorient_script_path: string, path to bash script to reorient image
         - robustfov_script_path: string, path to bash script to robustfov image
     '''
-    TMPDIR = os.path.join(preprocess_dir, "tmp_intermediate_preprocessing_steps")
-    if not os.path.exists(TMPDIR):
-        os.makedirs(TMPDIR)
+    TMPDIR = os.path.join(
+        preprocess_dir, "tmp_intermediate_preprocessing_steps")
 
     class_directories = [os.path.join(train_dir, x)
                          for x in os.listdir(train_dir)]
@@ -148,6 +174,8 @@ def preprocess_dir(train_dir, preprocess_dir, reorient_script_path, robustfov_sc
     # preprocess all images
     print("*** PREPROCESSING ***")
     for class_dir in tqdm(class_directories):
+        if not os.path.exists(TMPDIR):
+            os.makedirs(TMPDIR)
         preprocess_class_dir = os.path.join(
             preprocess_dir, os.path.basename(class_dir))
 
@@ -163,15 +191,19 @@ def preprocess_dir(train_dir, preprocess_dir, reorient_script_path, robustfov_sc
 
         # preprocess in parallel using all but one cores (n_jobs=-2)
         Parallel(n_jobs=ncores)(delayed(preprocess)(filename=f,
-                                                outdir=preprocess_class_dir,
-                                                tmpdir=TMPDIR,
-                                                reorient_script_path=reorient_script_path,
-                                                robustfov_script_path=robustfov_script_path,
-                                                verbose=0,)
-                            for f in filenames)
+                                                    outdir=preprocess_class_dir,
+                                                    tmpdir=TMPDIR,
+                                                    reorient_script_path=reorient_script_path,
+                                                    robustfov_script_path=robustfov_script_path,
+                                                    verbose=0,)
+                                for f in filenames)
 
-    # remove the intermediate preprocessing steps
-    shutil.rmtree(TMPDIR)
+        # remove the intermediate preprocessing steps at every iteration, otherwise
+        # disk usage goes beyond 100GB, with lots of training data
+        shutil.rmtree(TMPDIR)
+    # If the preprocessed data already exists, delete tmp_intermediate_preprocessing_steps
+    if os.path.exists(TMPDIR):
+        shutil.rmtree(TMPDIR)
 
 
 def load_image(filename):
@@ -183,20 +215,20 @@ def load_image(filename):
 def get_classes(task):
     class_encodings = {}
 
-    if task=="modality":
+    if task == "modality":
         class_encodings = {0: "FL",
                            1: "T1",
-                           2: "T2",}
-    elif task=="t1-contrast":
+                           2: "T2", }
+    elif task == "t1-contrast":
         class_encodings = {0: "T1 Post",
-                           1: "T1 Pre",}
-    elif task=="fl-contrast":
+                           1: "T1 Pre", }
+    elif task == "fl-contrast":
         class_encodings = {0: "FL Post",
-                           1: "FL Pre",}
+                           1: "FL Pre", }
     else:
         print("Invalid task: must be one of \"modality\", \"t1-contrast\", \"fl-contrast\"")
         sys.exit()
-        
+
     return class_encodings
 
 
@@ -219,19 +251,19 @@ def load_data(data_dir, labels_known=True):
         - all_filenames: list of strings, corresponding filenames for use in validation/test
     '''
 
-    data = []
     labels = []
-    all_filenames = []
 
     #################### CLASSIFICATION OF UNKNOWN DATA ####################
 
     if not labels_known:
+        all_filenames = []
         filenames = [x for x in os.listdir(data_dir)
                      if not os.path.isdir(os.path.join(data_dir, x))]
         filenames.sort()
 
-        for f in filenames:
+        for f in tqdm(filenames):
             img = nib.load(os.path.join(data_dir, f)).get_data()
+            img = np.reshape(img, img.shape+(1,))
             data.append(img)
             all_filenames.append(f)
 
@@ -245,6 +277,8 @@ def load_data(data_dir, labels_known=True):
     class_directories = [os.path.join(data_dir, x)
                          for x in os.listdir(data_dir)]
     class_directories.sort()
+    
+
     num_classes = len(class_directories)
 
     # point to the newly-processed files
@@ -252,19 +286,39 @@ def load_data(data_dir, labels_known=True):
                          for x in os.listdir(data_dir)]
     class_directories.sort()
 
-    for i in range(len(class_directories)):
-        filenames = os.listdir(class_directories[i])
-        filenames.sort()
+    # set up all_filenames and class_labels to speed up shuffling
+    all_filenames = []
+    class_labels = {}
+    i = 0
+    for class_directory in class_directories:
+        class_labels[os.path.basename(class_directory)] = i
+        i += 1
+        for filename in os.listdir(class_directory):
+            filepath = os.path.join(class_directory, filename)
+            all_filenames.append(filepath)
 
-        for f in filenames:
-            img = nib.load(os.path.join(class_directories[i], f)).get_data()
-            data.append(img)
-            labels.append(to_categorical(i, num_classes=num_classes))
-            all_filenames.append(f)
+    img_shape = nib.load(all_filenames[0]).get_data().shape
+    data = np.empty(shape=((len(all_filenames),) + img_shape + (1,)), dtype=np.uint8) 
 
-    data = np.array(data, dtype=np.float16)
-    labels = np.array(labels, dtype=np.float16)
+    # shuffle data
+    all_filenames = shuffle(all_filenames)
+        
+    data_idx = 0 # pointer to index in data
 
+    for f in tqdm(all_filenames):
+        img = nib.load(f).get_data()
+        img = np.asarray(img, dtype=np.uint8)
+        
+        # place this image in its spot in the data array
+        data[data_idx] = np.reshape(img, (1,)+img.shape+(1,))
+        data_idx += 1
+
+        cur_label = f.split(os.sep)[-2]
+        labels.append(to_categorical(class_labels[cur_label], num_classes=num_classes))
+
+    labels = np.array(labels, dtype=np.uint8)
+    print(data.shape)
+    print(labels.shape)
     return data, labels, all_filenames
